atualização botão logo e pagina de erro
