+++ processo-seletivo-compasso/src/pages/DetailsPage/styled.js
@@ -1,26 +1,9 @@
 import styled from 'styled-components';
-import { secondaryColor } from '../../constants/colors';
 
 export const DetailsContainer = styled.div `
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    padding-top: 5vh;
-`
-
-export const Logo = styled.img `
-    width: 80px;
-`
-
-export const LogoContainer = styled.div `
-    display: flex;
-    width: 20vw;
-    align-items: center;
-    justify-content: space-around;
-    margin-bottom: 3vh;
-    font-size: 40px;
-    font-weight: bold;
-    color: ${secondaryColor} ;
-    cursor: pointer;
+    padding-top: 8vh;
 `
