+++ processo-seletivo-compasso/src/pages/ErrorPage/ErrorPage.js
@@ -1,11 +1,14 @@
 import React from 'react';
+import LogoButton from '../../components/LogoButton/LogoButton';
+import { MainContainer } from './styled';
 
 const ErrorPage = () => {
 
     return (
-        <div>
-            ErrorPage
-        </div>
+        <MainContainer>
+            <LogoButton/>
+            <img src="https://media1.giphy.com/media/UHAYP0FxJOmFBuOiC2/giphy.gif" width="480" height="480" frameBorder="0" class="giphy-embed" allowFullScreen/>
+        </MainContainer>
     )
 }
