+++ processo-seletivo-compasso/src/pages/ErrorPage/styled.js
@@ -1,10 +1,9 @@
 import styled from 'styled-components';
 
-export const HomeContainer = styled.div `
+export const MainContainer = styled.div `
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     padding-top: 8vh;
 `
-
