+++ processo-seletivo-compasso/src/pages/HomePage/HomePage.js
@@ -1,16 +1,12 @@
 import React from 'react';
 import SearchField from '../../components/SearchField/SearchField';
-import { HomeContainer, Logo, LogoContainer } from './styled';
-import logo from '../../assets/logo.png';
-
+import { HomeContainer } from './styled';
+import LogoButton from '../../components/LogoButton/LogoButton';
 const HomePage = () => {
 
     return (
         <HomeContainer>
-            <LogoContainer>
-                <Logo src={logo} alt="logo"/>
-                GitUol
-            </LogoContainer>
+            <LogoButton/>
             <SearchField/>
         </HomeContainer>
     )
