+++ processo-seletivo-compasso/src/pages/SearchResultPage/SearchResultPage.js
@@ -1,27 +1,15 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import SearchField from '../../components/SearchField/SearchField';
 import SearchedCard from '../../components/SearchedCard/SearchedCard';
-import { HomeContainer, Logo, LogoContainer } from './styled';
-import logo from '../../assets/logo.png';
-import { useHistory } from 'react-router-dom';
-import { goToHome } from '../../routes/coordinator';
-import GlobalStateContext from '../../global/GlobalStateContext';
+import { HomeContainer } from './styled';
+import LogoButton from '../../components/LogoButton/LogoButton';
 
 const SearchResultPage = () => {
-    const history = useHistory()
-    const {setUser} = useContext(GlobalStateContext)
 
-    const onClickLogo = (history, setUser) => {
-        goToHome(history)
-        setUser(null)
-    }
 
     return (
         <HomeContainer>
-            <LogoContainer onClick={() => onClickLogo(history, setUser)}>
-                <Logo src={logo} alt="logo"/>
-                GitUol
-            </LogoContainer>
+            <LogoButton/>
             <SearchField/>
             <SearchedCard/>
         </HomeContainer>
